requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
Faker==30.3.0
idna==3.10
namecheap-python==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.5
rich==13.9.4
six==1.16.0
urllib3==2.5.0
